Add informer, fix bug
