+++ src/features/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 
 import News from 'features/news-page'
 import Profile from 'features/profile-page'
-import RequestMaterials from 'features/RequestMaterials/RequestMaterials'
+import RequestMaterials from 'features/RequestForm'
 
 import Admin, { Users, Events } from './admin'
 import Notifications from './notifications'
